dont need aliases if these exist

+++ cookbooks/aliases/files/default/admin.sh
@@ -1,7 +1,11 @@
 alias a1='ssh iad-admin101.ihr'
-alias chef='ssh use1b-ss-chef002.ihr'
-alias push-dns='ssh iad-admin101.ihr push-dns'B0;95;c
-alias consoleto='ssh iad-admin101.ihr -t sudo consoleto'
-alias rebooter='ssh iad-admin101.ihr -t sudo rebooter'
-alias poweron='ssh iad-admin101.ihr -t sudo poweron'
-alias poweroff='ssh iad-admin101.ihr -t sudo poweroff'
+if [ ! $(hostname) =~ "use1b-ss-chef002" ] ; then
+    alias chef='ssh use1b-ss-chef002.ihr'
+fi
+if [ ! $(hostname) =~ "iad-admin101.ihr" ] ; then
+    alias push-dns='ssh iad-admin101.ihr push-dns'
+    alias consoleto='ssh iad-admin101.ihr -t sudo consoleto'
+    alias rebooter='ssh iad-admin101.ihr -t sudo rebooter'
+    alias poweron='ssh iad-admin101.ihr -t sudo poweron'
+    alias poweroff='ssh iad-admin101.ihr -t sudo poweroff'
+fi
